requirements.txt
fastapi
uvicorn
asyncpg
httpx
python-dotenv
python-jose[cryptography]